+++ BlackFire/BFMessage.h
@@ -11,14 +11,14 @@
 @interface BFMessage : NSObject
 {
     NSString		*message; 
-    unsigned int	timestamp;
+    unsigned long	timestamp;
     unsigned int	user;
 }
 
-@property (nonatomic, assign) unsigned int timestamp;
+@property (nonatomic, assign) unsigned long timestamp;
 @property (nonatomic, assign) unsigned int user;
 @property (nonatomic, retain) NSString *message;
 
-- (id)initWithMessage:(NSString *)msg timestamp:(unsigned int)tstamp user:(unsigned int)usr;
+- (id)initWithMessage:(NSString *)msg timestamp:(unsigned long)tstamp user:(unsigned int)usr;
 
 @end
